DROOLS-5040: Minor refactoring
